Fixed a bug in the attribute handling:
This bug reveals itself when I was parsing a RELAX NG grammar with
RelaxNGCC. Here is the fragment:

      <element>
        <choice>
          <name>all</name>
          <name>choice</name>
          <name>sequence</name>
        </choice>
        <ref name="modelGroupBody" />
        ...

When startElement(ref) was encounered, the system was at
the NameClassState. Then it tries to revert to the "name" state.
But since a new attribute set that includes @name="modelGroupBody"
has already pushed as new attributes, the name state will wrongly
process the name attribute.

To fix this problem, I modified the way new Attributes is pushed.
Now Attributes are considered as a part of enterElement event, and
it is pushed when the event is actually consumed.

We consider it to be "actually consumed" when a state makes a transition
of enterElement. Oftentimes a state can spawn a child or revert to the parent,
but we don't consider it to be an actual consumption.