fix errors reported by John Cowan
----------------------------
> The documents are available at:
>
>   http://relaxngcc.sourceforge.net/workspace/relaxngcc/htdocs/en/index.htm

The manual (as opposed to the tutorials) does not mention c:class, which
made me doubtful what else it might be omitting.

Also, it says that the java-import element is like the java-import
attribute, which is not otherwise mentioned.  I think it should say
that the java-import element is like the java-body element.
----------------------------